kotlin2cpg: move file to different namespaces
request from @mpollmeier for scala3 migration
